Move ensemble_calibration to sit beneath statistical_postprocessing.
