+++ doc/source/extended_documentation/statistical_postprocessing/ensemble_calibration/create_coefficients_cube.rst
