+++ improver/ensemble_copula_coupling/ensemble_copula_coupling.py
@@ -40,7 +40,8 @@ from iris.exceptions import CoordinateNotFoundError, InvalidCubeError
 from scipy import stats
 
 from improver import BasePlugin
-from improver.ensemble_calibration.utilities import convert_cube_data_to_2d
+from improver.statistical_postprocessing.utilities import (
+    convert_cube_data_to_2d)
 from improver.ensemble_copula_coupling.utilities import (
     choose_set_of_percentiles, concatenate_2d_array_with_2d_array_endpoints,
     create_cube_with_percentiles, get_bounds_of_distribution,
@@ -634,8 +635,9 @@ class FromMeanAndVariance():
                 distribution constructed from the mean and standard deviation
                 provided.
                 Please note that for use with
-                :meth:`~improver.ensemble_calibration.ensemble_calibration.\
-ContinuousRankedProbabilityScoreMinimisers.calculate_truncated_normal_crps`,
+                :meth:`~improver.statistical_postprocessing.\
+ensemble_calibration.ContinuousRankedProbabilityScoreMinimisers.\
+calculate_truncated_normal_crps`,
                 the shape parameters for a truncated normal distribution with
                 a lower bound of zero should be [0, np.inf].
