+++ improver_tests/ensemble_copula_coupling/ensemble_copula_coupling/test_GeneratePercentilesFromProbabilities.py
@@ -46,11 +46,11 @@ from improver.ensemble_copula_coupling.ensemble_copula_coupling import \
 from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.utilities.warnings_handler import ManageWarnings
 
-from ...ensemble_calibration.ensemble_calibration.helper_functions import (
-    add_forecast_reference_time_and_forecast_period,
-    set_up_probability_above_threshold_spot_temperature_cube,
-    set_up_probability_above_threshold_temperature_cube,
-    set_up_probability_threshold_cube)
+from ...statistical_postprocessing.ensemble_calibration.helper_functions \
+    import (add_forecast_reference_time_and_forecast_period,
+            set_up_probability_above_threshold_spot_temperature_cube,
+            set_up_probability_above_threshold_temperature_cube,
+            set_up_probability_threshold_cube)
 
 
 class Test__add_bounds_to_thresholds_and_probabilities(IrisTest):
