+++ improver_tests/ensemble_copula_coupling/ensemble_copula_coupling/test_RebadgePercentilesAsMembers.py
@@ -44,8 +44,9 @@ from iris.tests import IrisTest
 from improver.ensemble_copula_coupling.ensemble_copula_coupling import \
     RebadgePercentilesAsRealizations as Plugin
 
-from ...ensemble_calibration.ensemble_calibration.helper_functions import (
-    add_forecast_reference_time_and_forecast_period, set_up_temperature_cube)
+from ...statistical_postprocessing.ensemble_calibration.helper_functions \
+    import (add_forecast_reference_time_and_forecast_period,
+            set_up_temperature_cube)
 
 
 class Test_process(IrisTest):
