+++ improver_tests/statistical_postprocessing/ensemble_calibration/test_ApplyCoefficientsForEnsembleCalibration.py
@@ -41,9 +41,9 @@ import numpy as np
 from iris.tests import IrisTest
 from numpy.testing import assert_array_almost_equal
 
-from improver.ensemble_calibration.ensemble_calibration import \
+from improver.statistical_postprocessing.ensemble_calibration import \
     ApplyCoefficientsFromEnsembleCalibration as Plugin
-from improver.ensemble_calibration.ensemble_calibration import (
+from improver.statistical_postprocessing.ensemble_calibration import (
     EstimateCoefficientsForEnsembleCalibration)
 from improver.metadata.constants.attributes import MANDATORY_ATTRIBUTE_DEFAULTS
 from improver.utilities.warnings_handler import ManageWarnings
