+++ improver_tests/statistical_postprocessing/ensemble_calibration/test_ContinuousRankedProbabilityScoreMinimisers.py
@@ -40,9 +40,10 @@ import iris
 import numpy as np
 from iris.tests import IrisTest
 
-from improver.ensemble_calibration.ensemble_calibration import \
+from improver.statistical_postprocessing.ensemble_calibration import \
     ContinuousRankedProbabilityScoreMinimisers as Plugin
-from improver.ensemble_calibration.utilities import convert_cube_data_to_2d
+from improver.statistical_postprocessing.utilities import (
+    convert_cube_data_to_2d)
 from improver.utilities.warnings_handler import ManageWarnings
 
 from .helper_functions import EnsembleCalibrationAssertions, SetupCubes
