+++ improver_tests/statistical_postprocessing/ensemble_calibration/test_EstimateCoefficientsForEnsembleCalibration.py
@@ -42,10 +42,10 @@ import iris
 import numpy as np
 from iris.tests import IrisTest
 
-from improver.ensemble_calibration.ensemble_calibration import (
+from improver.statistical_postprocessing.ensemble_calibration import (
     ContinuousRankedProbabilityScoreMinimisers)
-from improver.ensemble_calibration.ensemble_calibration import \
-    EstimateCoefficientsForEnsembleCalibration as Plugin
+from improver.statistical_postprocessing.ensemble_calibration import (
+    EstimateCoefficientsForEnsembleCalibration as Plugin)
 from improver.utilities.warnings_handler import ManageWarnings
 
 from ...set_up_test_cubes import set_up_variable_cube
@@ -283,7 +283,7 @@ class Test__repr__(IrisTest):
                "current_cycle: 20171110T0000Z; "
                "desired_units: None; "
                "predictor_of_mean_flag: mean; "
-               "minimiser: <class 'improver.ensemble_calibration."
+               "minimiser: <class 'improver.statistical_postprocessing."
                "ensemble_calibration."
                "ContinuousRankedProbabilityScoreMinimisers'>; "
                "coeff_names: ['gamma', 'delta', 'alpha', 'beta']; "
@@ -304,7 +304,7 @@ class Test__repr__(IrisTest):
                "current_cycle: 20171110T0000Z; "
                "desired_units: Kelvin; "
                "predictor_of_mean_flag: realizations; "
-               "minimiser: <class 'improver.ensemble_calibration."
+               "minimiser: <class 'improver.statistical_postprocessing."
                "ensemble_calibration."
                "ContinuousRankedProbabilityScoreMinimisers'>; "
                "coeff_names: ['gamma', 'delta', 'alpha', 'beta']; "
